opencl improvements

+++ src/include/openclStuff.h
@@ -4,4 +4,4 @@ void setup();
 
 double* cl_getResults(const double minA, const double maxA,
                       const double minB, const double maxB,
-                      const int step, const int numIterations);
+                      const uint64_t step, const int numIterations);
